+++ Commands/Base/SPOnlineConnectionHelper.cs
@@ -39,7 +39,7 @@ namespace OfficeDevPnP.PowerShell.Commands.Base
                 realm = GetRealmFromTargetUrl(url);
             }
 
-            var context = authManager.GetAppOnlyAuthenticatedContext(url.ToString(), realm, clientId, clientSecret);
+            var context = PnPClientContext.ConvertFrom(authManager.GetAppOnlyAuthenticatedContext(url.ToString(), realm, clientId, clientSecret), retryCount, retryWait * 1000);
             context.ApplicationName = Properties.Resources.ApplicationName;
             context.RequestTimeout = requestTimeout;
 
@@ -69,7 +69,7 @@ namespace OfficeDevPnP.PowerShell.Commands.Base
             string configFile = Path.Combine(appDataFolder, "OfficeDevPnP.PowerShell\\tokencache.dat");
             FileTokenCache cache = new FileTokenCache(configFile);
 
-            var context = authManager.GetAzureADNativeApplicationAuthenticatedContext(url.ToString(), clientId, redirectUri, cache);
+            var context = PnPClientContext.ConvertFrom(authManager.GetAzureADNativeApplicationAuthenticatedContext(url.ToString(), clientId, redirectUri, cache), retryCount, retryWait * 10000);
 
             var connectionType = ConnectionType.OnPrem;
             if (url.Host.ToUpperInvariant().EndsWith("SHAREPOINT.COM"))
@@ -89,8 +89,7 @@ namespace OfficeDevPnP.PowerShell.Commands.Base
         internal static SPOnlineConnection InitiateAzureADAppOnlyConnection(Uri url, string clientId, string tenant, string certificatePath, SecureString certificatePassword, int minimalHealthScore, int retryCount, int retryWait, int requestTimeout, bool skipAdminCheck = false)
         {
             Core.AuthenticationManager authManager = new Core.AuthenticationManager();
-            var context = authManager.GetAzureADAppOnlyAuthenticatedContext(url.ToString(), clientId, tenant, certificatePath, certificatePassword);
-
+            var context = PnPClientContext.ConvertFrom(authManager.GetAzureADAppOnlyAuthenticatedContext(url.ToString(), clientId, tenant, certificatePath, certificatePassword), retryCount, retryWait * 1000);
             var connectionType = ConnectionType.OnPrem;
             if (url.Host.ToUpperInvariant().EndsWith("SHAREPOINT.COM"))
             {
@@ -111,10 +110,12 @@ namespace OfficeDevPnP.PowerShell.Commands.Base
         {
             var authManager = new Core.AuthenticationManager();
 
-            var context = authManager.GetWebLoginClientContext(url.ToString());
+            var context = PnPClientContext.ConvertFrom(authManager.GetWebLoginClientContext(url.ToString()), retryCount, retryWait * 1000);
 
             if (context != null)
             {
+                context.RetryCount = retryCount;
+                context.Delay = retryWait * 1000;
                 context.ApplicationName = Properties.Resources.ApplicationName;
                 context.RequestTimeout = requestTimeout;
                 var connectionType = ConnectionType.OnPrem;
@@ -137,7 +138,9 @@ namespace OfficeDevPnP.PowerShell.Commands.Base
 
         internal static SPOnlineConnection InstantiateSPOnlineConnection(Uri url, PSCredential credentials, PSHost host, bool currentCredentials, int minimalHealthScore, int retryCount, int retryWait, int requestTimeout, bool skipAdminCheck = false)
         {
-            ClientContext context = new ClientContext(url.AbsoluteUri);
+            var context = new PnPClientContext(url.AbsoluteUri);
+            context.RetryCount = retryCount;
+            context.Delay = retryWait * 1000;
             context.ApplicationName = Properties.Resources.ApplicationName;
             context.RequestTimeout = requestTimeout;
             if (!currentCredentials)
@@ -215,7 +218,9 @@ namespace OfficeDevPnP.PowerShell.Commands.Base
                 throw new Exception("Cannot retrieve ADFS settings.");
             }
 
-            var context = authManager.GetADFSUserNameMixedAuthenticatedContext(url.ToString(), networkCredentials.UserName, networkCredentials.Password, networkCredentials.Domain, adfsHost, adfsRelyingParty);
+            var context = PnPClientContext.ConvertFrom(authManager.GetADFSUserNameMixedAuthenticatedContext(url.ToString(), networkCredentials.UserName, networkCredentials.Password, networkCredentials.Domain, adfsHost, adfsRelyingParty), retryCount, retryWait * 1000);
+            context.RetryCount = retryCount;
+            context.Delay = retryWait * 1000;
 
             context.ApplicationName = Properties.Resources.ApplicationName;
             context.RequestTimeout = requestTimeout;
@@ -302,7 +307,8 @@ namespace OfficeDevPnP.PowerShell.Commands.Base
                         adfsRelyingParty = queryParameters["wtrealm"];
                     }
                 }
-            } catch(WebException ex)
+            }
+            catch (WebException ex)
             {
                 throw new Exception("Endpoint does not use ADFS for authentication.", ex);
             }