+++ Commands/ContentTypes/SetDefaultContentTypeToList.cs
@@ -10,8 +10,9 @@ namespace OfficeDevPnP.PowerShell.Commands
     [CmdletHelp("Sets the default content type for a list", 
         Category = CmdletHelpCategory.ContentTypes)]
     [CmdletExample(
-     Code = @"PS:> Set-SPODefaultContentTypeToList -List ""Project Documents"" -ContentType ""Project""",
-     Remarks = @"This will set the Project content type (which has already been added to a list) as the default content type", SortOrder = 1)]
+        Code = @"PS:> Set-SPODefaultContentTypeToList -List ""Project Documents"" -ContentType ""Project""",
+        Remarks = @"This will set the Project content type (which has already been added to a list) as the default content type", 
+        SortOrder = 1)]
     public class SetDefaultContentTypeToList : SPOWebCmdlet
     {
         [Parameter(Mandatory = true)]