+++ Commands/Extensions/WebExtensions.cs
@@ -11,9 +11,7 @@ namespace OfficeDevPnP.PowerShell.Commands
             var clientContext = currentWeb.Context as ClientContext;
             Site site = clientContext.Site;
             Web web = site.OpenWebById(guid);
-            clientContext.Load(web, w => w.Url, w => w.Title, w => w.Id);
-            clientContext.ExecuteQueryRetry();
-
+            web.EnsureProperties(w => w.Url, w => w.Title, w => w.Id, w => w.ServerRelativeUrl);
             return web;
         }
 
@@ -22,9 +20,7 @@ namespace OfficeDevPnP.PowerShell.Commands
             var clientContext = currentWeb.Context as ClientContext;
             Site site = clientContext.Site;
             Web web = site.OpenWeb(url);
-            clientContext.Load(web, w => w.Url, w => w.Title, w => w.Id);
-            clientContext.ExecuteQueryRetry();
-
+            web.EnsureProperties(w => w.Url, w => w.Title, w => w.Id, w => w.ServerRelativeUrl);
             return web;
         }