+++ Commands/WebParts/GetWebPart.cs
@@ -4,6 +4,7 @@ using System;
 using System.Linq;
 using System.Management.Automation;
 using OfficeDevPnP.PowerShell.CmdletHelpAttributes;
+using OfficeDevPnP.Core.Utilities;
 
 namespace OfficeDevPnP.PowerShell.Commands
 {