+++ changelog.md
@@ -1,5 +1,20 @@
 # OfficeDevPnP.PowerShell Changelog #
 
+**2016-04-08**
+* Added -ExtensibilityHandlers parameter to Get-SPOPRovisioningTemplate
+
+**2016-03-11**
+* Added List parameter to Get-SPOContentType, allowing to retrieve the ContentTypes added to a list.
+
+**2016-03-08**
+* Added Remove-SPOListItem
+* Updated Get-SPOWeb and Get-SPOSubWebs to include ServerRelativeUrl
+* Added Ensure-SPOFolder cmdlet
+
+**2016-03-07**
+* Added Remove-SPOFieldFromContentType cmdlet
+* Added Get-SPOSiteSearchQueryResults cmdlet
+
 **2016-02-04**
 * Added -PersistPublishingFiles and -IncludeNativePublishingFiles parameters to Get-SPOProvisioningTemplate